+++ src/main/java/domain/Lines.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.List;
