+++ src/main/java/utils/exception/ExceptionHandler.java
@@ -1,6 +1,7 @@
-package utils;
+package utils.exception;
 
-import org.slf4j.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ExceptionHandler {
     private static final Logger logger = LoggerFactory.getLogger(ExceptionHandler.class);
