+++ src/main/java/utils/generator/ReservedPositionGenerator.java
@@ -1,7 +1,11 @@
-package model;
+package utils.generator;
 
-import java.util.*;
-import java.util.stream.*;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Random;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class ReservedPositionGenerator {
     private static final Random RANDOM = new Random();
